Update Problem51
After searching for numbers ranging from 10**1 to 10**9, went online to find help.
My old code had glanced over the answer without reporting it which is weird considering that I did not change isPrime(x) or prime8Check(x)
Whatever. I used information online that directed me to only view numbers that had three repeats, this removes the likelihood of having guaranteed multiples of 3 which takes away the possibility of 8 primes out of 10.
Zero failed so I moved on to One
Found other numbers that factored in three 0's in the front that were technically also prime numbers.
Finished, and I do not want to review because I am still annoyed at the original method not detecting the solution for some bizarre reason.