Update Problem68.py
Added restrict()
Because we were only looking for the maximum 16 digit number sequence, we could remove all cases where 10 was an inner node. 
I spent the whole time trying to find the largest sequence which I assumed would be a 17 digit number, thus doing the exact opposite of what was required.